+++ crates/gamemaster/src/devserver.rs
@@ -15,7 +15,7 @@ pub struct DevPlayerState {
     pub id: String,
     pub name: String,
     pub current_speed_kmh: f32,
-    pub total_distance_m: i32,
+    pub total_distance_m: f64,
     pub is_walking: bool,
     pub map_tile_x: i32,
     pub map_tile_y: i32,
@@ -193,7 +193,7 @@ fn handle_request(request: &str, state: &SharedState, events: &SharedEvents, not
             struct WalkerReq {
                 player_id: String,
                 speed: f32,
-                distance: i32,
+                distance: f64,
                 #[serde(default)]
                 incline: f32,
                 #[serde(default)]
